Changed seat type from long to int

+++ BusBookingSystem/src/main/java/com/booking/system/model/Bus.java
@@ -29,7 +29,7 @@ public class Bus {
     private LocalDateTime departureTime;
     private LocalDateTime arrivalTime;
     private LocalTime travelTime;
-    private List<Long> availableSeats;
+    private List<Integer> availableSeats;
     private long ticketPrice;
 
     public void setTravelTime() {
